# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/kegiatan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan daftar semua kegiatan dengan filter dan pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegiatan"
                ],
                "summary": "Mendapatkan semua kegiatan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per halaman (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter berdasarkan ID kategori",
                        "name": "kategori_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal mulai (format: YYYY-MM-DD)",
                        "name": "tanggal_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter tanggal sampai (format: YYYY-MM-DD)",
                        "name": "tanggal_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat kegiatan baru dengan data yang diberikan",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegiatan"
                ],
                "summary": "Membuat kegiatan baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama kegiatan",
                        "name": "kegiatan_nama",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID kategori kegiatan",
                        "name": "kategori_kegiatan_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tanggal kegiatan (format: RFC3339)",
                        "name": "kegiatan_tanggal",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lokasi kegiatan",
                        "name": "kegiatan_lokasi",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Penanggung jawab kegiatan",
                        "name": "kegiatan_pj",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Deskripsi kegiatan",
                        "name": "kegiatan_deskripsi",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kegiatan berhasil dibuat",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal membuat kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/kegiatan/bulan/{bulan}/tahun/{tahun}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan daftar kegiatan berdasarkan bulan dan tahun tertentu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegiatan"
                ],
                "summary": "Mendapatkan kegiatan per bulan dan tahun",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bulan (1-12)",
                        "name": "bulan",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Tahun (2000-2100)",
                        "name": "tahun",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kegiatan bulanan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parameter tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/kegiatan/mendatang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan daftar kegiatan yang akan datang (tanggal \u003e= hari ini)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegiatan"
                ],
                "summary": "Mendapatkan kegiatan mendatang",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jumlah data (default: 5)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kegiatan mendatang",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/kegiatan/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mencari kegiatan berdasarkan kata kunci",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegiatan"
                ],
                "summary": "Mencari kegiatan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kata kunci pencarian",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pencarian kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parameter pencarian harus diisi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Gagal mencari kegiatan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
}